Add support for encoding and decoding smaller int types
and adapt code generation to use the new int types when appropriate.

This is adapted from changes by Dinne Bosman (@dwjbosman)

+++ tests/encode/test3_corner_cases/src/main.c
@@ -1039,8 +1039,8 @@ ZTEST(cbor_encode_test3, test_single)
 	};
 	size_t input_single1 = 52;
 	size_t input_single2_ign = 53;
-	size_t input_single3 = 9;
-	size_t input_single4_inv = 10;
+	uint8_t input_single3 = 9;
+	uint8_t input_single4_inv = 10;
 
 	zassert_equal(ZCBOR_SUCCESS, cbor_encode_SingleBstr(output, sizeof(output), &input_single0, &out_len), NULL);
 	zassert_equal(sizeof(exp_payload_single0), out_len, NULL);
